Разделил VC на расширения и фалы по MVC
