+++ TeskTaskProject/TeskTaskProject/Extension/ExtensionColorView.swift
@@ -1,0 +1,17 @@
+//
+//  ExtensionColorView.swift
+//  TeskTaskProject
+//
+//  Created by Александр Косяков on 15.11.2022.
+//
+
+import UIKit
+
+extension MainViewController {
+    
+    func colorView () {
+        view.backgroundColor = UIColor(red: 237/255, green: 231/255, blue: 229/255, alpha: 1)
+        headerView.backgroundColor = UIColor(red: 237/255, green: 231/255, blue: 229/255, alpha: 1)
+        footerView.backgroundColor = UIColor(red: 237/255, green: 231/255, blue: 229/255, alpha: 1)
+    }
+}
